requirements.txt
Django~=3.2.12
django-fsm==2.8.0
djangorestframework==3.13.1
djangorestframework-jsonp==1.0.2
django-pagedown==2.2.1
django-markdown-deux
django-registration==3.2
psycopg2==2.8.4
Pillow==6.2.1
requests
gunicorn[gevent]==20.1.0
gevent==21.12.0
profanityfilter
jinja2==3.0.3
markupsafe==2.0.1
django-cors-headers
numpy==1.21.2
git+https://github.com/astropy/astropy
astroplan==0.8
astroscrappy==1.1.0
fits2image>=0.4.0
fits_align>=0.4.3
boto3
django-storages
